+++ spec/lib/castle/request_spec.rb
@@ -1,19 +1,16 @@
 # frozen_string_literal: true
 
 describe Castle::Request do
-  subject do
-    described_class.new(headers)
-  end
+  subject { described_class.new(headers) }
 
   let(:headers) { { 'SAMPLE-HEADER' => '1' } }
   let(:api_secret) { Castle.config.api_secret }
 
   describe 'build' do
-    let(:path) { 'endpoint' }
-    let(:params) { { user_id: 1 } }
-
     context 'get' do
-      let(:request) { subject.build_query(path) }
+      let(:path) { 'endpoint' }
+      let(:params) { { user_id: 1 } }
+      let(:request) { subject.build(path, params, :get) }
 
       it { expect(request.body).to be_nil }
       it { expect(request.method).to eql('GET') }
@@ -23,6 +20,8 @@ describe Castle::Request do
     end
 
     context 'post' do
+      let(:path) { 'endpoint' }
+      let(:params) { { user_id: 1 } }
       let(:request) { subject.build(path, params, :post) }
 
       it { expect(request.body).to be_eql('{"user_id":1}') }