+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - `Castle::Client` now takes options as a second argument
 - drop support for ruby 2.1
 - replaced `config.api_endpoint` with `config.host` and `config.port`
+- renamed `fetch_review` to `Castle::Review.retrieve`
 
 **Features:**